git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@10190 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/lightmap.h
@@ -63,7 +63,7 @@ class CLightmapFaceData {
 
 class CLightmapData : public CLightmapFaceData {
 	public:
-		int						nOffset [MAX_LIGHTMAP_WIDTH];
+		float						nOffset [MAX_LIGHTMAP_WIDTH];
 		CArray<tSegFacePtr>	faceList;
 		int						nBlackLightmaps;
 		int						nWhiteLightmaps;
